detection_tags: split demo data synthesis out of the pipeline
Move all COCO/demo-only code (constants, download+cache, gamma darkening) into a dedicated
detection/coco_demo.py (CocoDemoSource + darken). steps.load_batch is now source-agnostic
plumbing: it pulls (bytes, boxes, labels) from the source and lays them into the pipeline
tables. For real data you replace one file (coco_demo.py); steps.py/app.py are untouched.
